ci-daemon: route per-job nice into Choir weight
job_nice() in jobs.rs computes per-test-group nice plus historical-stats
adjustments and stores the result in Job.nice — but make_job_spec was
only reading rc.ktest.user_nice, so the per-test-group and per-stats
nice values were dropped on the floor. Cross-branch scheduling didn't
reflect group-level priorities because branches mapping to different
test_groups all looked the same to Choir's weight.

Sum job.nice and user_nice at the use site. Widen Job.nice to i64 so it
combines with i64 user_nice without sign games.

user_nice fairness across users is a separate concern — leaving its
current weight-additive behavior in place for now.

+++ src/jobs.rs
@@ -44,7 +44,7 @@ pub struct Job {
     /// Commits behind the branch tip (0 = tip). Lower = run sooner.
     pub age: u64,
     /// Niceness — higher = lower scheduling priority.
-    pub nice: u64,
+    pub nice: i64,
     /// Expected runtime in seconds, from historical durations.
     pub duration: u64,
 }
@@ -128,17 +128,17 @@ fn job_wanted(status: Option<TestStatus>) -> bool {
 /// Niceness for one subtest: the test_group's base nice plus the
 /// historical-stats adjustments — nice down tests that consistently
 /// pass-or-fail, and long-running tests.
-fn job_nice(tg: &RcTestGroup, stats: Option<&TestStats>) -> u64 {
-    let mut nice = tg.nice;
+fn job_nice(tg: &RcTestGroup, stats: Option<&TestStats>) -> i64 {
+    let mut nice = tg.nice as i64;
     if let Some(stats) = stats {
         if tg.test_always_passes_nice != 0
             && stats.passed != stats.failed
             && stats.passed + stats.failed > tg.test_always_passes_nice
         {
-            nice += tg.test_always_passes_nice;
+            nice += tg.test_always_passes_nice as i64;
         }
         if tg.test_duration_nice != 0 {
-            nice += stats.duration / tg.test_duration_nice;
+            nice += (stats.duration / tg.test_duration_nice) as i64;
         }
     }
     nice
